# dbmongo/docs/swagger/swagger.yaml
components:
  schemas:
    Login:
      properties:
        email:
          type: string
        password:
          type: string
        browserToken: 
          type: string
    LoginResponse:
      properties:
        code:
          type: integer
        expire:
          type: string
        token:
          type: string
    LoginCheck:
      properties:
        email:
          type: string
        password:
          type: string
        checkCode: 
          type: string
    LoginCheckResponse:
      properties:
        browserToken:
          type: string
    LoginGet:
      properties:
        email:
          type: string
        password:
          type: string
    LoginRecoveryGet:
      properties:
        email:
          type: string
        browsertoken:
          type: string
    LoginRecoveryGetResponse: 
      type: string
...
    LoginFail:
      properties:
        code:
          type: integer
        message:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
info:
  contact: 
    name: "contact support"
    email: "contact@signaux-faibles.beta.gouv.fr"
  description: |-
    Cette API centralise toutes les fonctionnalités du module de traitement de données [openSignauxFaibles](https://github.com/entrepreneur-interet-general/opensignauxfaibles)  
    Ce [projet](https://beta.gouv.fr/startups/signaux-faibles.html) fournit une plateforme technique à l'exploitation des signaux faibles pour anticiper les difficultés des entreprises  
  
    La plateforme fournit les services suivants    
    - Authentification  
    - Gestion des batches d'intégration  
    - Lancement/ordonnancement des traitements  
    - Consultation des données  
    
    Toutes les erreurs rencontrées lors du traitement seront signalées par le statut `HTTP 500` assorti du message d'erreur ayant provoqué le défaut du traitement.  
    Par ailleurs, tous les services authentifiés sont succeptibles de renvoyer le statut `HTTP 401` dans le cas où l'authentification est erronée.  Le code `HTTP 403` est également communiqué lorsque les attributions sont insuffisantes pour traiter la requête.  
    Dans le cas où une requête est malformée, le code `HTTP 400` sera renvoyé assorti d'un message indiquant l'erreur rencontrée dans l'analyse du corps de la requête  
  license:
    name: Licence MIT
    url: https://raw.githubusercontent.com/entrepreneur-interet-general/opensignauxfaibles/master/LICENSE
  title: API openSignauxFaibles
  version: "1.1"
paths:
  /login:
    post:
      description: |-
        L'authentification se base sur la fourniture de 3 champs
        - email
        - mot de passe 
        - jeton du navigateur

        Le service retourne un jeton temporaire valide pour une durée déterminée.  
        Cette validité peut être prolongée avec le service /api/refreshToken  
        
        Ce token doit être utilisé dans l'entête `Authorization`
      requestBody:
        description: Les informations d'identification
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        "200":
          description: retourne un token et une date de validité
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                code: 200
                expire: "2018-12-31 23:30:12"
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9…"
        "401":
          description: retourne la raison de l'échec
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginFail'
              example:
                code: 401
                message: "Please provide good credentials sir."
      summary: obtenir un jeton temporaire de session
      tags:
      - Authentification
  /login/get:
    post:
      summary: "demander l'envoi d'un code de vérification"
      description: |-
        Renvoie un code de vérification sur la base d'un email et d'un mot de passe  
        Le code n'est envoyé que si le mot de passe est valide, le cas échéant, un email avertissant d'une tentative est envoyé.  
        Pour éviter les tentatives de forçage de mot de passe, ce service ne renvoie jamais d'échec.  
      requestBody:
        description: Email et mot de passe
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginGet'
      responses:
        "200":
          description: "ok"
      tags:
      - Authentification
  /login/check:
    post:
      description: |-
        Le service permet de valider le code de vérification envoyé à l'utilisateur  
        Il donne en retour un jeton de navigateur qui pourra être réutilisé comme facteur d'authentification (browserToken)  
      requestBody:
        description: Les informations d'identification
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCheck'
      responses:
        "200":
          description: retourne un browserToken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginCheckResponse'
              example:
                browserToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9…"
      summary: valider un code de vérification
      tags:
      - Authentification
  /login/recovery/get:
    post:
      description: |-
        Dans le cas d'un oubli du mot de passe, on peut fixer un nouveau mot de passe à partir d'un navigateur identifié  
        Il faut dans ce cas pouvoir recevoir un code de vérification par mail  
        Ce schéma d'authentification se base sur la fourniture d'un browserToken valide associé à l'utilisateur avec lequel il a été obtenu ainsi qu'un code de vérification transmis par email
      requestBody:
        description: Les informations d'identification
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRecoveryGet'
      responses:
        "200":
          description: retourne un browserToken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginRecoveryGetResponse'
              example:
                email: "test@domain.com"
                browserToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9…"
      summary: demander une récupération de mot de passe
      tags:
      - Authentification
...
  /api/refreshToken:
    get:
      summary: renouveler le jeton de session
      description: |-
        Fournit un nouveau jeton temporaire de session en échange d'un jeton encore valide.  
        Ce service demande d'être authentifié mais ne nécessite aucun paramètre.  
        L'authentification est réalisée avec l'entête HTTP Authorization.  
        Exemple:  
        ```Authorization:Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9…```
      responses:
        "200":
          description: Renouvellement autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                code: 200
                expire: "2019-02-01 12:00:05"
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9…"
        "401":
          description: Renouvellement refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginFail'
              example:
                code: 401
                message: "cookie token is empty"
      security:
      - BearerAuth: []
      tags:
      - Authentification
...
openapi: "3.0.0"